requirements.txt
aiogram==3.6.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
python-dotenv==1.0.1
requests